chore: Apply SOLID - DIP
