+++ app/lib/repositories/analysis/analysis_repository_impl.dart
@@ -1,0 +1,64 @@
+import 'package:earth_and_i/domains/type/e_challenge.dart';
+import 'package:earth_and_i/domains/type/e_user_status.dart';
+import 'package:earth_and_i/providers/analysis/analysis_provider.dart';
+import 'package:earth_and_i/repositories/analysis/analysis_repository.dart';
+import 'package:get/get.dart';
+
+class AnalysisRepositoryImpl extends GetxService implements AnalysisRepository {
+  late final AnalysisProvider _analysisProvider;
+
+  @override
+  void onInit() {
+    super.onInit();
+    _analysisProvider = Get.find<AnalysisProvider>();
+  }
+
+  @override
+  Future<Map<String, dynamic>> analysisAction(
+    EUserStatus userStatus,
+    String question,
+    String answer,
+  ) async {
+    Map<String, dynamic> result;
+
+    try {
+      result = await _analysisProvider.postAnalysisText(
+        userStatus,
+        question,
+        answer,
+      );
+    } catch (e) {
+      rethrow;
+    }
+
+    bool isGood = result["carbon"]["reduce"] == " good";
+    double changeCapacity = double.parse(result["carbon"]["output"]);
+
+    return {
+      "answer": result["answer"],
+      "changeCapacity": isGood ? -changeCapacity : changeCapacity,
+    };
+  }
+
+  @override
+  Future<Map<String, dynamic>> analysisChallenge(
+    EChallenge challenge,
+    String image,
+  ) async {
+    Map<String, dynamic> result;
+
+    try {
+      result = await _analysisProvider.postAnalysisImage(
+        challenge,
+        image,
+      );
+    } catch (e) {
+      rethrow;
+    }
+
+    return {
+      "type": result["type"],
+      "res": result["res"] == "True",
+    };
+  }
+}
