+++ app/lib/view_models/setting/setting_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:earth_and_i/models/setting/alarm_state.dart';
-import 'package:earth_and_i/repositories/action_history/action_history_repository.dart';
+import 'package:earth_and_i/repositories/action_history/action_history_repository_impl.dart';
 import 'package:earth_and_i/repositories/challenge_history/challenge_history_repository.dart';
 import 'package:earth_and_i/repositories/user/user_repository.dart';
 import 'package:earth_and_i/utilities/functions/security_util.dart';
@@ -15,7 +15,7 @@ class SettingViewModel extends GetxController {
   /* -------------------- DI Fields ----------------------- */
   /* ------------------------------------------------------ */
   late final UserRepository _userRepository;
-  late final ActionHistoryRepository _actionHistoryRepository;
+  late final ActionHistoryRepositoryImpl _actionHistoryRepository;
   late final ChallengeHistoryRepository _challengeHistoryRepository;
 
   /* ------------------------------------------------------ */
@@ -37,7 +37,7 @@ class SettingViewModel extends GetxController {
     super.onInit();
     // DI Fields
     _userRepository = Get.find<UserRepository>();
-    _actionHistoryRepository = Get.find<ActionHistoryRepository>();
+    _actionHistoryRepository = Get.find<ActionHistoryRepositoryImpl>();
     _challengeHistoryRepository = Get.find<ChallengeHistoryRepository>();
 
     // Private Fields
